requirements.txt
fastapi
uvicorn
mediapipe
opencv-python
numpy
pillow